refactor: #184 Footer 컴포넌트를 components/common 로 이동
