--- src/components/common/Footer/FooterStyle.ts
@@ -1,22 +1,0 @@
-import styled from '@emotion/styled';
-
-export const FooterContainer = styled.div`
-  min-width: 320px;
-  max-width: 425px;
-  height: 8vh;
-  display: flex;
-  position: fixed;
-  left: 0;
-  bottom: 0;
-  width: 100%;
-  z-index: 777;
-`;
-
-export const FooterButton = styled.button`
-  cursor: pointer;
-  width: 100%;
-  height: 100%;
-  border-radius: 0;
-  border: 1px solid;
-  background-color: white;
-`;
