+++ src/components/common/Footer/assets/index.ts
